requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
pymongo[srv]
dnspython
